+++ app/src/scenes/onboarding-v2/checkIn/OnboardingCheckInStartScreen.tsx
@@ -1,26 +1,40 @@
-import React from "react";
+import React, { useCallback } from "react";
 import { Text } from "react-native";
 import { OnboardingV2ScreenProps } from "../types";
 import BeigeWrapperScreen from "../BeigeWrapperScreen";
 import BeigeCard from "../BeigeCard";
 import { mergeClassNames } from "@/utils/className";
 import { typography } from "@/utils/typography";
-import { VARIANT_BORDER_COLORS } from "../data/carouselData";
+import logEvents from "@/services/logEvents";
+import { VARIANT_BORDER_COLORS, VARIANT_COLORS } from "../data/carouselData";
+import { useFocusEffect } from "@react-navigation/native";
+import { setStatusBarBackgroundColor } from "expo-status-bar";
+import { TW_COLORS } from "@/utils/constants";
+import { useStatusBar } from "@/context/StatusBarContext";
 type Props = OnboardingV2ScreenProps<"Intro">;
 
 const NextRoute = "OnboardingCheckInHowDoYouFeel";
 
 export const OnboardingCheckInStartScreen: React.FC<Props> = ({ navigation }) => {
+  const { setCustomColor } = useStatusBar();
+
+  useFocusEffect(
+    useCallback(() => {
+      setCustomColor("#E5F6FC");
+      return () => {
+        // Optional cleanup here
+      };
+    }, [])
+  );
+
   const handleNext = () => {
+    logEvents.logHumeurObdStart();
     navigation.navigate(NextRoute);
   };
 
   const handlePrevious = () => {
+    logEvents.logOnboardingBack(7);
     navigation.goBack();
-  };
-
-  const handleSkip = () => {
-    handleNext();
   };
 
   return (
@@ -29,6 +43,7 @@ export const OnboardingCheckInStartScreen: React.FC<Props> = ({ navigation }) =>
       nextText="Commencer ma première observation"
       handlePrevious={handlePrevious}
       handleNext={handleNext}
+      variant="blue"
     >
       <BeigeCard color={VARIANT_BORDER_COLORS.blue}>
         <Text className={mergeClassNames(typography.displayXsRegular, "text-left mb-6 text-cnam-primary-900")}>